Read version one location
